[#104] Rename: move WorkoutCard, SelectableWorkoutCard to cards folder
- SelectWorkout, SetupResult, NewPlan 페이지에서 components/cards 풀더 내 WorkoutCard, SelectableCard를 사용하도록 수정함
